+++ backpack-compose/src/main/kotlin/net/skyscanner/backpack/compose/calendar2/internal/BpkCalendarSpaceCell.kt
@@ -1,0 +1,41 @@
+/*
+ * Backpack for Android - Skyscanner's Design System
+ *
+ * Copyright 2018 Skyscanner Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.skyscanner.backpack.compose.calendar2.internal
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import net.skyscanner.backpack.calendar2.data.CalendarCell
+import net.skyscanner.backpack.compose.utils.applyIf
+
+@Composable
+internal fun BpkCalendarSpaceCell(
+  model: CalendarCell.Space,
+  modifier: Modifier = Modifier,
+) {
+  Spacer(
+    modifier = modifier.applyIf(model.selected) {
+      background(
+        brush = CalendarDayBackground.selectionTopBrush(CalendarCell.Selection.Middle),
+        shape = CalendarDayBackground.selectionTopShape(CalendarCell.Selection.Middle),
+      )
+    }
+  )
+}
